py-telegram-bot/requirements.txt
dotenv
python-telegram-bot
